+++ website/app.js
@@ -1,7 +1,6 @@
 const express = require("express");
 const app = express();
 const dotenv = require("dotenv");
-const mariadb = require("mariadb");
 
 // controllers
 const AccountController = require("./controllers/Account");
@@ -14,18 +13,12 @@ var { buildSchema } = require("graphql");
 dotenv.config();
 
 const port = process.env.PORT || 3000;
-
-// Connect to database
-const pool = mariadb.createPool({
-  host: process.env.DB_HOST,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASS,
-  database: process.env.DB_NAME,
-  port: process.env.DB_PORT,
-  connectionLimit: 5,
-});
+const pool = require("./models/Database");
 
 Account = new AccountController(pool);
+
+// routers
+const IndexRouter = require("./routes/Index");
 
 // Test database connection
 pool
@@ -40,37 +33,10 @@ pool
   });
 
 // Construct a schema, using GraphQL schema language
-var schema = buildSchema(`
-
-    type AccountType {
-        accountID: Int,
-        username: String
-        geboortedatum: String,
-        aanmelddatum: String
-    }
-
-    type Query {
-        accounts: String,
-        account(username: String): AccountType
-        login(username: String, password: String): Boolean
-    }        
-`);
-
+var schema = require("./models/Schema");
 
 // The root provides a resolver function for each API endpoint
-var root = {
-  accounts: async () => {
-    let accounts = await AccountModel.getAccounts();
-    return accounts[0].username;
-  },
-  account: async ({ username }) => {
-
-    return await Account.getAccount(username);
-  },
-    login: async ({ username, password }) => {
-        return await Account.login(username, password);
-    }
-};
+var root = require("./models/Root");
 
 app.use(
   "/graphql",
@@ -81,26 +47,7 @@ app.use(
   })
 );
 
-app.get("/", (req, res) => {
-  let users = [];
-  pool
-    .getConnection()
-    .then((conn) => {
-      conn
-        .query("SELECT * FROM accounts")
-        .then((rows) => {
-          users = rows;
-          conn.release();
-          res.send(users);
-        })
-        .catch((err) => {
-          console.log("Error querying database");
-        });
-    })
-    .catch((err) => {
-      console.log("Error connecting to database");
-    });
-});
+app.use("/", IndexRouter);
 
 app.listen(port, () => {
   console.log(`Listening on http://localhost:${port}`);
